refactor(routing_actor): consolidate into single platform-independent file
